+++ example/js/test_build_in_module.js
@@ -1,1 +1,0 @@
-var buildInModule = require('build_in_module');