+++ example/main.js
@@ -1,3 +1,5 @@
-document.getElementById("msg").innerHTML = "hello, Cube";
-document.getElementById("show").value = document.getElementById("initscript").innerHTML;
-exports.run = function () {console.log("app started!")};
+require('./tests');
+
+exports.run = function () {
+  console.log('app start');
+};