+++ test/index.js
@@ -9,89 +9,93 @@ var testMod = require('../index');
 var path = require('path');
 var request = require('supertest');
 testMod.init({
-  root: path.join(__dirname, '../example') + '/',
+  root: path.join(__dirname, '../example'),
   port: 7777,
-  router: '/a/b/',
+  router: '/',
   middleware: false,
-  buildInModule: {
-    'build_in_module': true
-  }
 });
 request = request('http://localhost:7777');
+
 describe('index.js', function () {
   describe('query js files', function () {
     it('should return regular js file', function (done) {
-      request.get('/a/b/js/jquery.js')
+      request.get('/main.js')
         .expect(200)
         .expect('content-type', 'application/javascript')
+        .expect(/exports\.run/)
         .expect(/^(?!Cube\()/, done);
     });
     it('should return regular js file which name with dot', function (done) {
-      request.get('/a/b/js/file.dot.js')
+      request.get('/test/test_file.with.dot.js')
         .expect(200)
         .expect('content-type', 'application/javascript')
         .expect(/^(?!Cube\()/, done);
     });
+    it('should return regular js file which name with hyphen', function (done) {
+      request.get('/test/test_file-with-hyphen.js?m')
+        .expect(200)
+        .expect('content-type', 'application/javascript')
+        .expect(/^Cube\(/, done);
+    });
     it('should return transfered js file', function (done) {
-      request.get('/a/b/js/index.js?m')
+      request.get('/main.js?m')
         .expect(200)
         .expect('content-type', 'application/javascript')
         .expect(function (res) {
           var body = res.text;
-          expect(body).to.match(/require\('\/js\/jquery\.js'\)/);
-          expect(body).to.match(/Cube\("\/js\/index\.js"/);
+          expect(body).to.match(/require\('\/tests\.js'\)/);
+          expect(body).to.match(/Cube\("\/main\.js"/);
         })
         .end(done);
     });
     it('should return transfered coffee file', function (done) {
-      request.get('/a/b/js/test_coffee.coffee?m')
+      request.get('/test/test_coffee.coffee?m')
         .expect(200)
         .expect('content-type', 'application/javascript')
         .expect(function (res) {
           var body = res.text;
-          expect(body).to.match(/exports.run\s+=/);
-          expect(body).to.match(/Cube\("\/js\/test_coffee\.coffee"/);
+          expect(body).to.match(/hello coffee-script/);
+          expect(body).to.match(/Cube\("\/test\/test_coffee\.coffee"/);
         })
         .end(done);
     });
     it('should return transfered coffee file call like a js file', function (done) {
-      request.get('/a/b/js/test_coffee.js?m')
+      request.get('/test/test_coffee.js?m')
         .expect(200)
         .expect('content-type', 'application/javascript')
         .expect(function (res) {
           var body = res.text;
-          expect(body).to.match(/exports.run\s+=/);
-          expect(body).to.match(/Cube\("\/js\/test_coffee\.js"/);
+          expect(body).to.match(/hello coffee-script/);
+          expect(body).to.match(/Cube\("\/test\/test_coffee\.js"/);
         })
         .end(done);
     });
 
     it('should return a compress file', function (done) {
-      request.get('/a/b/js/index.js?m&c')
+      request.get('/test/test_main.js?m&c')
         .expect(200)
         .expect('content-type', 'application/javascript')
         .expect(function (res) {
           var body = res.text;
-          expect(body).to.match(/Cube\("\/js\/index\.js/);
+          expect(body).to.match(/Cube\("\/test\/test_main\.js/);
           expect(body).not.match(/\/\*\!/);
-          expect(body).not.match(/module/);
         })
         .end(done);
     });
     it('should return err message when file not found when moduleWrap is on', function (done) {
-      request.get('/a/b/js/jquery-notfound.js?m')
+      request.get('/test/module-not-found.js?m')
         .expect(200)
         .expect(/console\.error/)
         .expect(/file not found/)
-        .expect(/jquery-notfound/, done);
+        .expect(/module-not-found/, done);
     });
     it('should return err message when file not found when normal query', function (done) {
-      request.get('/a/b/js/jquery-notfound.js')
+      request.get('/test/module-not-found.js')
         .expect(404)
         .expect(/file not found/, done);
     });
     it('should return error message in console when file parse ast error', function (done) {
-      request.get('/a/b/js/error.js?m')
+      request.get('/test/test_error.js?m')
         .expect(200)
         .expect(function (res) {
           expect(res.text).match(/JS_Parser_Error/ig);
@@ -99,7 +103,7 @@ describe('index.js', function () {
         .end(done);
     });
     it('should return ok when file require node_modules', function (done) {
-      request.get('/a/b/js/node_modules_test.js?m')
+      request.get('/test/test_module.js?m')
         .expect(200)
         .expect(function (res) {
           expect(res.text).to.match(/\/node_modules\/test\/lib\/a\.js/);
@@ -108,19 +112,11 @@ describe('index.js', function () {
         })
         .end(done);
     });
-    it('should return 404 when file require node_modules not exist', function (done) {
-      request.get('/a/b/js/node_modules_not_found.js?m')
+    it('should return 200 when require node_modules not exist, they maybe registered in the page', function (done) {
+      request.get('/test/test_registered_module.js?m')
         .expect(200)
         .expect(function (res) {
-          expect(res.text).to.match(/required module not found/);
-        })
-        .end(done);
-    });
-    it('should ignore build in modules', function (done) {
-      request.get('/a/b/js/test_build_in_module.js?m')
-        .expect(200)
-        .expect(function (res) {
-          expect(res.text).to.match(/\["build_in_module"\]/);
+          expect(res.text).to.match(/\["jquery"\]/);
         })
         .end(done);
     });
@@ -128,7 +124,7 @@ describe('index.js', function () {
 
   describe('query css file', function () {
     it('should return a transfered css file', function (done) {
-      request.get('/a/b/css/test.css')
+      request.get('/css/test_css.css')
         .expect(200)
         .expect(function (res) {
           expect(res.text).match(/\.test \{/ig);
@@ -136,44 +132,63 @@ describe('index.js', function () {
         .end(done);
     });
     it('should return a transfered comressed css file', function (done) {
-      request.get('/a/b/css/test.css?c')
+      request.get('/css/test_css.css?c')
         .expect(200)
         .expect(function (res) {
           expect(res.text).match(/\.test\{/ig);
         })
         .end(done);
     });
-    it('should return a transfered comressed css file with wrap, actually a js module', function (done) {
-      request.get('/a/b/css/test.css?m&c')
+    it('should return a wraped & comressed css file, actually a js module', function (done) {
+      request.get('/css/test_css.css?m&c')
         .expect(200)
         .expect('content-type', 'application/javascript')
         .expect(function (res) {
           expect(res.text).match(/\.test\{/ig);
-          expect(res.text).match(/^Cube\("\/css\/test\.css", *\[\]/);
-          expect(res.text).match(/\.test\{color:#f30\}/);
+          expect(res.text).match(/^Cube\("\/css\/test_css\.css", *\[\]/);
+          expect(res.text).match(/\.test\{color:/);
         })
         .end(done);
     });
-    it('should return a transfered less file', function (done) {
-      request.get('/a/b/css/test_less.less')
+    it('should return the source less file', function (done) {
+      request.get('/css/test_less.less')
         .expect(200)
         .expect(function (res) {
-          expect(res.text).match(/\.box a \{/ig);
+          expect(res.text).match(/\@red/ig);
+          expect(res.text).not.match(/\.test \.box/ig);
+        })
+        .end(done);
+    });
+    it('should return the transfered less file', function (done) {
+      request.get('/css/test_less.css')
+        .expect(200)
+        .expect('content-type', 'text/css')
+        .expect(function (res) {
+          expect(res.text).match(/\.test \.box/ig);
         })
         .end(done);
     });
     it('should return a transfered compressed less file', function (done) {
-      request.get('/a/b/css/test_less.less?c')
+      request.get('/css/test_less.less?c')
         .expect(200)
         .expect(function (res) {
-          expect(res.text).match(/\.box a\{/ig);
+          expect(res.text).match(/\.test \.box a\{/ig);
           expect(res.text).not.match(/\n/);
+        })
+        .end(done);
+    });
+    it('should return a transfered & compressed & wraped less file', function (done) {
+      request.get('/css/test_less.less?m&c')
+        .expect(200)
+        .expect(function (res) {
+          expect(res.text).match(/\.test \.box a\{/ig);
+          expect(res.text).match(/^Cube\("\/css\/test_less\.less"/);
         })
         .end(done);
     });
 
     it('should return a transfered styl file', function (done) {
-      request.get('/a/b/css/test_styl.styl?m')
+      request.get('/css/test_styl.styl?m')
         .expect(200)
         .expect(function (res) {
           expect(res.text).match(/\.test a \{/ig);
@@ -181,8 +196,8 @@ describe('index.js', function () {
         .end(done);
     });
 
-    it('should return a transfered styl file', function (done) {
-      request.get('/a/b/css/test_styl.styl?m&c')
+    it('should return a transfered compressed styl file', function (done) {
+      request.get('/css/test_styl.styl?m&c')
         .expect(200)
         .expect(function (res) {
           expect(res.text).match(/\.test a\{/ig);
@@ -191,20 +206,41 @@ describe('index.js', function () {
         .end(done);
     });
 
-    it('shoud seekFile success', function (done) {
-      request.get('/a/b/css/test.styl?c')
+    it('shoud return a compressed css file, but not transfered', function (done) {
+      request.get('/css/test_styl.css?c')
         .expect(200)
         .expect(function (res) {
-          expect(res.text).match(/\.test\{color:#f30\}/ig);
+          expect(res.text).match(/\.test a\{color:#f30\}/ig);
           expect(res.text).not.match(/\n/);
+          expect(res.text).not.match(/^Cube/);
         })
         .end(done);
     });
   });
 
   describe('query tpl file', function () {
+    it('should return a regular html file', function (done) {
+      request.get('/tpl/test.html')
+        .expect(200)
+        .expect('content-type', 'text/html')
+        .expect(function (res) {
+          var code = res.text;
+          expect(code).not.match(/^Cube\(/);
+          expect(code).to.match(/<h3>/);
+        }).end(done);
+    });
+    it('should return a wraped html file', function (done) {
+      request.get('/tpl/test.html?m')
+        .expect(200)
+        .expect('content-type', 'application/javascript')
+        .expect(function (res) {
+          var code = res.text;
+          expect(code).match(/^Cube\(/);
+          expect(code).to.match(/<h3>/);
+        }).end(done);
+    });
     it('should return a compiled ejs file', function (done) {
-      request.get('/a/b/tpl/test_ejs.ejs?m')
+      request.get('/tpl/test_ejs.ejs?m')
         .expect(200)
         .expect(function (res) {
           var code = res.text;
@@ -212,7 +248,7 @@ describe('index.js', function () {
         }).end(done);
     });
     it('should return a compiled jade file', function (done) {
-      request.get('/a/b/tpl/test_jade.jade?m')
+      request.get('/tpl/test_jade.jade?m')
         .expect(200)
         .expect(function (res) {
           var code = res.text;