+++ tools.js
@@ -1,4 +1,3 @@
-var G = require('./global');
 var xfs = require('xfs');
 var path = require('path');
 var ug = require('uglify-js');
@@ -43,7 +42,7 @@ function checkIgnore(file, ignores) {
   return flag;
 }
 
-function processDir(source, dest, options, cb) {
+function processDir(cube, source, dest, options, cb) {
   if (!dest) {
     return console.log('[ERROR] param missing! dest');
   }
@@ -75,57 +74,46 @@ function processDir(source, dest, options, cb) {
       console.log('[copy file]:', relFile.substr(1));
       return;
     }
-    var type =  G.processors.map[ext];
+    var type =  cube.processors.map[ext];
     if (type === undefined) {
       // unknow type, copy file
       xfs.sync().save(destFile, xfs.readFileSync(sourceFile));
       console.log('[copy file]:', relFile.substr(1));
       return;
     }
-    var ps = G.processors.types[type];
+    var ps = cube.processors.types[type];
     var processor = ps[ext];
-    // lazy loading processing
-    if (typeof processor === 'string') {
-      try {
-        processor = require(processor);
-        G.processors.types[type][ext] = processor;
-      } catch (e) {
-        e.message = 'loading transform error: type:`' + type + '` ext:`' + ext + '`';
-        e.code = 'CUBE_LOADING_TRANSFORM_ERROR';
-        errors.push(e);
-      }
-    }
     var options = {
       moduleWrap: true,
       sourceMap: false,
       compress: true,
-      buildInModule: G.buildInModule,
-      release: true
+      release: cube.config.release,
+      root: cube.config.root,
+      qpath: relFile,
     };
-    processor(source, relFile, options, function (err, result) {
+    processor.process(relFile, options, function (err, result) {
       if (err) {
         console.log('[ERROR]', err.message);
         return errors.push(err);
       }
-      var finalFile, wrapDestFile, wrapCode, modName;
+      var finalFile, wrapDestFile, modName;
       if (type === 'script') {
         destFile = destFile.replace(/\.\w+$/, '.js');
         var destSourceFile = destFile.replace(/\.js/, '.source.js');
-        xfs.sync().save(destFile, result.min);
+        xfs.sync().save(destFile, result.code);
         xfs.sync().save(destSourceFile, result.source);
         console.log('[transfer script]:', relFile.substr(1));
       } else if (type === 'style') {
         finalFile = destFile.replace(/\w+$/, 'css');
         wrapDestFile = destFile + '.js';
         modName = relFile + '.js';
-        wrapCode = 'Cube("' + modName + '",[],function(){return ' + JSON.stringify(result.min) + '});';
-        xfs.sync().save(wrapDestFile, wrapCode);
-        xfs.sync().save(finalFile, result.min);
+        xfs.sync().save(wrapDestFile, result.wraped);
+        xfs.sync().save(finalFile, result.code);
         console.log('[transfer style]:', relFile.substr(1));
       } else if (type === 'template') {
         wrapDestFile = destFile + '.js';
         xfs.sync().save(destFile, result.source);
-        xfs.sync().save(wrapDestFile, result.wrap);
+        xfs.sync().save(wrapDestFile, result.wraped);
         console.log('[transfer template]:', relFile.substr(1));
       }
     });